refactor: src 및 src/view 파일을 TypeScript로 변환
