+++ src/registerElement.ts
